Add support for MIRACLE LINUX
The installer for RPM-based systems refuses to run on MIRACLE LINUX,
which is a RHEL-derivative and compatible with node-red.

This commit adds MIRACLE LINUX to the list of recognized distributions
of the installer for RPM-based systems.

I have confirmed that this works on MIRACLE LINUX 8.4.

```
Running nodejs and Node-RED install for user asdf at /home/asdf on "miraclelinux"

  Stop Node-RED                       ✔
  Install Node.js LTS                 ✔   Node v14.19.0   Npm 6.14.16
  Install Node-RED core               ✔   2.2.0
  Add shortcut commands               ✔
  Update systemd script               ✔
  Update public zone firewall rule    ✔

Any errors will be logged to   /var/log/nodered-install.log
All done.
  You can now start Node-RED with the command  node-red-start
  Then point your browser to localhost:1880 or http://{your_ip-address}:1880

Started  2022年  2月  2日 水曜日 16:22:47 JST  -  Finished  2022年  2月  2日 水曜日 16:23:21 JST
```